refactor(docs): move texts components to their own dir
