+++ src/pages/ui-components/Heading/index.tsx
@@ -4,7 +4,7 @@ import { IStyledProp } from '@eduzz/houston-styles/styled';
 import useHoustonTheme from '@eduzz/houston-styles/useHoustonTheme';
 import type { HoustonTokens } from '@eduzz/houston-tokens';
 
-import useMediaQuery from '../../hooks/useMediaQuery';
+import useMediaQuery from '../hooks/useMediaQuery';
 import Typography, { TypographyColors } from '../Typography';
 
 type HeadingTags = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6';
