+++ src/pages/ui-components/Subtitle/index.tsx
@@ -4,7 +4,7 @@ import { IStyledProp } from '@eduzz/houston-styles/styled';
 import useHoustonTheme from '@eduzz/houston-styles/useHoustonTheme';
 import type { HoustonTokens } from '@eduzz/houston-tokens';
 
-import useMediaQuery from '../../hooks/useMediaQuery';
+import useMediaQuery from '../hooks/useMediaQuery';
 import Typography, { TypographyColors } from '../Typography';
 
 type SubtitleTags = 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'p';
@@ -13,7 +13,7 @@ export type SubtitleSizes = 'large' | 'small';
 
 export type SubtitleColors = TypographyColors;
 
-export interface IHeadingProps extends IStyledProp {
+export interface ISubtitleProps extends IStyledProp {
   id?: string;
   /**
    * Defaults to 'large'
@@ -49,7 +49,7 @@ const mobileSizesMap: SizesMap = {
   small: 'sm'
 };
 
-const Subtitle = React.forwardRef<HTMLHeadingElement | HTMLParagraphElement, IHeadingProps>(
+const Subtitle = React.forwardRef<HTMLHeadingElement | HTMLParagraphElement, ISubtitleProps>(
   ({ as = 'h2', children, size: sizeProp = 'large', ...props }, ref) => {
     const { breakpoints } = useHoustonTheme();
     const isMobile = useMediaQuery(breakpoints.down('sm'));
